Extract feed card sub components
